+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotEngineImpl.java
@@ -239,6 +239,8 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
 
     final APIAccess apiAccess;
 
+    final boolean probeAssertionsEnabled;
+
     @SuppressWarnings("unchecked")
     PolyglotEngineImpl(PolyglotImpl impl, SandboxPolicy sandboxPolicy, String[] permittedLanguages,
                     DispatchOutputStream out, DispatchOutputStream err, InputStream in, OptionValuesImpl engineOptions,
@@ -257,6 +259,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
         this.in = in;
         this.logHandler = logHandler;
         this.logLevels = logLevels;
+        this.probeAssertionsEnabled = initAssertProbes(engineOptions);
         this.hostLanguage = createLanguage(LanguageCache.createHostLanguageCache(hostImpl), PolyglotEngineImpl.HOST_LANGUAGE_INDEX, null);
         this.boundEngine = boundEngine;
         this.storeEngine = RUNTIME.isStoreEnabled(engineOptions);
@@ -351,6 +354,17 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
                 deprecatedDescriptors.add(d);
             }
         }
+    }
+
+    private static boolean initAssertProbes(OptionValuesImpl engineOptions) {
+        boolean assertsOn = false;
+        assert !!(assertsOn = true);
+        boolean assertProbes = engineOptions.get(PolyglotEngineOptions.AssertProbes);
+        if (assertProbes && !assertsOn) {
+            throw PolyglotEngineException.illegalState("Option engine.AssertProbes is set to true, but assertions are disabled. " +
+                            "Assertions need to be enabled for this option to be functional. Pass -ea as VM option to resolve this problem.\n");
+        }
+        return assertProbes;
     }
 
     /**
@@ -489,6 +503,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
         this.host = prototype.host;
         this.boundEngine = prototype.boundEngine;
         this.logHandler = prototype.logHandler;
+        this.probeAssertionsEnabled = prototype.probeAssertionsEnabled;
         this.hostLanguage = createLanguage(LanguageCache.createHostLanguageCache(prototype.getHostLanguageSPI()), HOST_LANGUAGE_INDEX, null);
         this.engineLoggerSupplier = prototype.engineLoggerSupplier;
 
@@ -2045,7 +2060,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
         boolean enterReverted = false;
         if (CompilerDirectives.injectBranchProbability(CompilerDirectives.LIKELY_PROBABILITY, info.getThread() == Thread.currentThread())) {
             // Volatile increment is safe if only one thread does it.
-            prev = info.enterInternal();
+            prev = info.enterInternal(this);
 
             // Check again whether the cached thread info is still the same as expected
             if (CompilerDirectives.injectBranchProbability(CompilerDirectives.FASTPATH_PROBABILITY, info == context.getCachedThread())) {
@@ -2058,7 +2073,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
                 try {
                     info.notifyEnter(this, context);
                 } catch (Throwable e) {
-                    info.leaveInternal(prev);
+                    info.leaveInternal(this, prev);
                     throw e;
                 }
                 return prev;
@@ -2067,7 +2082,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
                  * If we go this path and enteredCount drops to 0, the subsequent slowpath enter
                  * must call deactivateThread.
                  */
-                info.leaveInternal(prev);
+                info.leaveInternal(this, prev);
                 enterReverted = true;
             }
         }
@@ -2118,7 +2133,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
             try {
                 info.notifyLeave(this, context);
             } finally {
-                info.leaveInternal(prev);
+                info.leaveInternal(this, prev);
                 entered = false;
             }
             if (CompilerDirectives.injectBranchProbability(CompilerDirectives.FASTPATH_PROBABILITY, info == context.getCachedThread())) {